move file related logic to internal/files
