+++ internal/files/config.go
@@ -1,21 +1,21 @@
-package config
+package files
 
 type fileConfig struct {
 	Path     string
 	JQFilter string
 }
 
-var AllSets fileConfig = fileConfig{
+var allSets fileConfig = fileConfig{
 	Path:     "./src/SetList.json",
 	JQFilter: "{meta, data: [.data | .[] | {code, name}]}",
 }
 
-var AllAtomicCards fileConfig = fileConfig{
+var allAtomicCards fileConfig = fileConfig{
 	Path:     "./src/AtomicCards.json",
 	JQFilter: "{meta, data: [.data | to_entries | .[] | .value | .[0] | {name, identifiers}]}",
 }
 
-var AllSetCards fileConfig = fileConfig{
+var allSetCards fileConfig = fileConfig{
 	Path:     "./src/AllIdentifiers.json",
 	JQFilter: "{meta, data: [.data | to_entries | .[] | .value | {uuid, identifiers, name, setCode, number}]}",
 }
